ref: Move server logic to handler functions
