+++ cmd/server/main.go
@@ -2,8 +2,6 @@ package main
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
 	"log/slog"
 	"net"
 	"net/http"
@@ -12,122 +10,19 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/1saswata/chess-broadcast-engine/internal/auth"
+	"github.com/1saswata/chess-broadcast-engine/internal/api"
 	"github.com/1saswata/chess-broadcast-engine/internal/broker"
 	"github.com/1saswata/chess-broadcast-engine/internal/cache"
 	"github.com/1saswata/chess-broadcast-engine/internal/db"
 	"github.com/1saswata/chess-broadcast-engine/internal/pb"
 	"github.com/1saswata/chess-broadcast-engine/internal/server"
 	"github.com/1saswata/chess-broadcast-engine/internal/telemetry"
-	"github.com/google/uuid"
 
 	"google.golang.org/grpc"
 )
 
 type UserHandler struct {
 	db.UserRepository
-}
-
-func login(uh UserHandler, rc *cache.RedisCache) {
-	mux := http.NewServeMux()
-	mux.HandleFunc("POST /register", func(w http.ResponseWriter, r *http.Request) {
-		v := struct {
-			Username string `json:"username"`
-			Password string `json:"password"`
-			Role     string `json:"role"`
-		}{}
-		err := json.NewDecoder(r.Body).Decode(&v)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-		hPass, err := auth.HashPassword(v.Password)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-		ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
-		defer cancel()
-		err = uh.CreateUser(ctx, v.Username, hPass, v.Role)
-		if err != nil {
-			if err == context.DeadlineExceeded {
-				http.Error(w, "Database is slow", http.StatusGatewayTimeout)
-				return
-			}
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-		w.WriteHeader(http.StatusCreated)
-	})
-	mux.HandleFunc("POST /login", func(w http.ResponseWriter, r *http.Request) {
-		v := struct {
-			Username string `json:"username"`
-			Password string `json:"password"`
-			MatchID  int32  `json:"match_id"`
-		}{}
-		w.Header().Set("Access-Control-Allow-Origin", "*")
-		err := json.NewDecoder(r.Body).Decode(&v)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-		ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
-		defer cancel()
-		user, err := uh.GetUserByUsername(ctx, v.Username)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusUnauthorized)
-			return
-		}
-		if !auth.CheckPasswordHash(v.Password, user.PasswordHash) {
-			http.Error(w, "Wrong password", http.StatusUnauthorized)
-			return
-		}
-		token, err := auth.GenerateToken(user.ID, v.MatchID, user.Role)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-		fmt.Fprintf(w, "%s", token)
-	})
-	mux.HandleFunc("POST /archive", func(w http.ResponseWriter, r *http.Request) {
-		v := struct {
-			MatchID       int32     `json:"match_id"`
-			WhitePlayerId uuid.UUID `json:"white_player_id"`
-			BlackPlayerID uuid.UUID `json:"black_player_id"`
-		}{}
-		err := json.NewDecoder(r.Body).Decode(&v)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-		ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
-		defer cancel()
-		moveHistory, err := rc.GetMoveHistory(ctx, v.MatchID)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		err = uh.ArchiveMatch(ctx, v.MatchID, v.WhitePlayerId, v.BlackPlayerID,
-			moveHistory)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		key := fmt.Sprintf("match:%d:latest", v.MatchID)
-		err = rc.DeleteKey(ctx, key)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		key = fmt.Sprintf("match:%d:sequence", v.MatchID)
-		err = rc.DeleteKey(ctx, key)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-	})
-	newServer := http.Server{Addr: ":8080", Handler: mux}
-	slog.Error("Error on login", "Error", newServer.ListenAndServe())
 }
 
 func main() {
@@ -169,7 +64,15 @@ func main() {
 		slog.Error("Error connecting to cache", "Error", err)
 		os.Exit(1)
 	}
-	go login(uh, rc)
+	apiServer := api.NewAPIServer(uh.UserRepository, rc)
+	mux := apiServer.SetupRouter()
+	httpServer := http.Server{Addr: ":8080", Handler: mux}
+	go func() {
+		if err := httpServer.ListenAndServe(); err != nil &&
+			err != http.ErrServerClosed {
+			slog.Error("Http Server errpr", "Error", err)
+		}
+	}()
 	ingestServer := server.NewIngestServer(rp, rc)
 	lis, err := net.Listen("tcp", ":50051")
 	if err != nil {
